AI/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
ultralytics==8.3.176